requirements.txt
streamlit==1.40.0
pillow==11.0.0
pycryptodome==3.21.0
numpy==2.1.3
opencv-python-headless==4.8.0.76
